# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restoran va Kafe CRM/POS tizimi API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  owner: 'owner',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  /** @nullable */
  name?: string | null;
  role: UserRole;
  /** @nullable */
  venueId?: number | null;
  /** @nullable */
  venueName?: string | null;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  owner: 'owner',
  admin: 'admin',
} as const;

export interface UserInput {
  username: string;
  password: string;
  name?: string;
  role: UserInputRole;
  /** @nullable */
  venueId?: number | null;
}

export type VenueType = typeof VenueType[keyof typeof VenueType];


export const VenueType = {
  cafe: 'cafe',
  restaurant: 'restaurant',
} as const;

export interface Venue {
  id: number;
  name: string;
  type: VenueType;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  adminId?: number | null;
  /** @nullable */
  adminName?: string | null;
  createdAt?: string;
}

export type VenueInputType = typeof VenueInputType[keyof typeof VenueInputType];


export const VenueInputType = {
  cafe: 'cafe',
  restaurant: 'restaurant',
} as const;

export interface VenueInput {
  name: string;
  type: VenueInputType;
  address?: string;
  phone?: string;
}

export type VenueUpdateType = typeof VenueUpdateType[keyof typeof VenueUpdateType];


export const VenueUpdateType = {
  cafe: 'cafe',
  restaurant: 'restaurant',
} as const;

export interface VenueUpdate {
  name?: string;
  type?: VenueUpdateType;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
}

export interface AdminAssignInput {
  userId: number;
}

export interface VenueStats {
  todaySales: number;
  totalRevenue: number;
  totalDebts: number;
  productCount: number;
  orderCount: number;
}

export interface Product {
  id: number;
  venueId: number;
  name: string;
  price: number;
  category: string;
  /** @nullable */
  description?: string | null;
  isAvailable?: boolean;
  createdAt?: string;
}

export interface ProductInput {
  name: string;
  price: number;
  category: string;
  description?: string;
  isAvailable?: boolean;
}

export interface ProductUpdate {
  name?: string;
  price?: number;
  category?: string;
  /** @nullable */
  description?: string | null;
  isAvailable?: boolean;
}

export interface Customer {
  id: number;
  venueId: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  totalDebt?: number;
  createdAt?: string;
}

export interface CustomerInput {
  name: string;
  phone?: string;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  total: number;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export type OrderPaymentType = typeof OrderPaymentType[keyof typeof OrderPaymentType];


export const OrderPaymentType = {
  cash: 'cash',
  debt: 'debt',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  completed: 'completed',
  debt: 'debt',
} as const;

export interface Order {
  id: number;
  venueId: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  totalAmount: number;
  paymentType: OrderPaymentType;
  status: OrderStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type OrderDetailPaymentType = typeof OrderDetailPaymentType[keyof typeof OrderDetailPaymentType];


export const OrderDetailPaymentType = {
  cash: 'cash',
  debt: 'debt',
} as const;

export type OrderDetailStatus = typeof OrderDetailStatus[keyof typeof OrderDetailStatus];


export const OrderDetailStatus = {
  completed: 'completed',
  debt: 'debt',
} as const;

export interface OrderDetail {
  id: number;
  venueId: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  totalAmount: number;
  paymentType: OrderDetailPaymentType;
  status: OrderDetailStatus;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  createdAt: string;
}

export type OrderInputPaymentType = typeof OrderInputPaymentType[keyof typeof OrderInputPaymentType];


export const OrderInputPaymentType = {
  cash: 'cash',
  debt: 'debt',
} as const;

export interface OrderInput {
  /** @nullable */
  customerId?: number | null;
  items: OrderItemInput[];
  paymentType: OrderInputPaymentType;
  notes?: string;
}

export type DebtStatus = typeof DebtStatus[keyof typeof DebtStatus];


export const DebtStatus = {
  unpaid: 'unpaid',
  paid: 'paid',
  partial: 'partial',
} as const;

export interface Debt {
  id: number;
  venueId: number;
  customerId: number;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  orderId: number;
  amount: number;
  paidAmount?: number;
  remaining?: number;
  status: DebtStatus;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}

export interface DebtPayInput {
  amount: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalSold: number;
  revenue: number;
}

export interface DashboardSummary {
  todayRevenue: number;
  todayOrderCount: number;
  totalDebt: number;
  unpaidDebtCount: number;
  topProducts: TopProduct[];
  recentOrders: Order[];
}

export interface VenueStat {
  venueId: number;
  venueName: string;
  todayRevenue: number;
  totalDebt: number;
  orderCount?: number;
}

export interface OwnerSummary {
  totalVenues: number;
  totalRevenue: number;
  totalDebt: number;
  venueStats: VenueStat[];
}