# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restoran va Kafe CRM/POS tizimi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Tizimga kirish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Muvaffaqiyatli kirish
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Noto'g'ri ma'lumotlar
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Joriy foydalanuvchi ma'lumotlari
      responses:
        "200":
          description: Foydalanuvchi ma'lumotlari
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Autentifikatsiya talab etiladi
...
  /venues/{venueId}/customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: Kafe mijozlari ro'yxati
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mijozlar ro'yxati
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Yangi mijoz qo'shish
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Mijoz yaratildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /venues/{venueId}/customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Mijoz ma'lumotlari
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mijoz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /venues/{venueId}/orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: Kafe buyurtmalari ro'yxati
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Buyurtmalar ro'yxati
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Yangi buyurtma (savdo)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Buyurtma yaratildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /venues/{venueId}/orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Buyurtma tafsilotlari (chek uchun)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Buyurtma tafsilotlari
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /venues/{venueId}/debts:
    get:
      operationId: listDebts
      tags: [debts]
      summary: Qarz daftar ro'yxati
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Qarzlar ro'yxati
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debt"
  /venues/{venueId}/debts/{id}/pay:
    patch:
      operationId: payDebt
      tags: [debts]
      summary: Qarzni to'langan deb belgilash
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtPayInput"
      responses:
        "200":
          description: Yangilandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
...
  /users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: Barcha foydalanuvchilar (owner uchun)
      responses:
        "200":
          description: Foydalanuvchilar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [auth]
      summary: Yangi foydalanuvchi yaratish (owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Yaratildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, admin]
        venueId:
          type: ["integer", "null"]
        venueName:
          type: ["string", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [owner, admin]
        venueId:
          type: ["integer", "null"]
...
    Customer:
      type: object
      required: [id, venueId, name]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        totalDebt:
          type: number
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice, total]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        total:
          type: number
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    Order:
      type: object
      required: [id, venueId, totalAmount, paymentType, status, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        paymentType:
          type: string
          enum: [cash, debt]
        status:
          type: string
          enum: [completed, debt]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderDetail:
      type: object
      required: [id, venueId, totalAmount, paymentType, status, items, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        paymentType:
          type: string
          enum: [cash, debt]
        status:
          type: string
          enum: [completed, debt]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [items, paymentType]
      properties:
        customerId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        paymentType:
          type: string
          enum: [cash, debt]
        notes:
          type: string
    Debt:
      type: object
      required: [id, venueId, customerId, customerName, orderId, amount, status, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        orderId:
          type: integer
        amount:
          type: number
        paidAmount:
          type: number
        remaining:
          type: number
        status:
          type: string
          enum: [unpaid, paid, partial]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
    DebtPayInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
...