# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restoran va Kafe CRM/POS tizimi API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentType } from './orderPaymentType';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  venueId: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  totalAmount: number;
  paymentType: OrderPaymentType;
  status: OrderStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}