# lib/api-zod/src/generated/types/orderDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restoran va Kafe CRM/POS tizimi API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDetailPaymentType } from './orderDetailPaymentType';
import type { OrderDetailStatus } from './orderDetailStatus';
import type { OrderItem } from './orderItem';

export interface OrderDetail {
  id: number;
  venueId: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  totalAmount: number;
  paymentType: OrderDetailPaymentType;
  status: OrderDetailStatus;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  createdAt: string;
}