# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restoran va Kafe CRM/POS tizimi API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputPaymentType } from './orderInputPaymentType';
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  /** @nullable */
  customerId?: number | null;
  items: OrderItemInput[];
  paymentType: OrderInputPaymentType;
  notes?: string;
}